Refactor LLM connection module
- Remove the old llm_connect.py file.
- Introduce new structure with separate files for each LLM provider (OpenAI, Anthropic, Gemini, etc.).
- Add __init__.py for module initialization.
- Implement available models and token limits in a dedicated file.
