+++ biochatter/llm_connect/__init__.py
@@ -1,0 +1,21 @@
+"""Init of the LLM connect module."""
+
+from biochatter.llm_connect.anthropic import AnthropicConversation
+from biochatter.llm_connect.azure import AzureGptConversation
+from biochatter.llm_connect.conversation import Conversation
+from biochatter.llm_connect.gemini import GeminiConversation
+from biochatter.llm_connect.misc import WasmConversation
+from biochatter.llm_connect.ollama import OllamaConversation
+from biochatter.llm_connect.openai import GptConversation
+from biochatter.llm_connect.xinference import XinferenceConversation
+
+__all__ = [
+    "AnthropicConversation",
+    "AzureGptConversation",
+    "Conversation",
+    "GeminiConversation",
+    "GptConversation",
+    "OllamaConversation",
+    "WasmConversation",
+    "XinferenceConversation",
+]
